Fixing Accelerate example
In the example by NyxOuf referenced : https://gist.github.com/Nyx0uf/217d97f81f4889f4445a

The CGSize is passed into the function which is then used to set the destination width and height. In the example here, the scalingFactor is undefined and multiplying by the image.size.width & height doesn't make sense to me in this context. I've declared the variable above to model off the example a bit more.