Fix pbcopy example in simctl article
I may be missing something, but I'm pretty sure the only way to use `pbcopy` is to pipe something into it, e.g. `cat ~/Desktop/file.txt | pbcopy`. Running the command in the article, `pbcopy ~/Desktop/file.txt` waits for more input on the command line.

Side note: I'm also confused by the definitions of `pbcopy` and `pbpaste` in the article. The article says "`pbpaste` takes the pasteboard contents of the Simulator and copies them to the desktop's main pasteboard. However, the example clearly does the opposite: takes the contents of the desktop's main pasteboard and copies it to the Simulator's pasteboard. The example is correct; this is how it works. Your description also matches Apple's documentation when you run `xcrun simctl pbpaste --help`. However, the docs don't seem to be in sync with the example.