Update “Never” article reference to Result type
At the time of the article Swift didn’t have a standard Result type, I changed the wording to reflect that it now does but didn’t at the time of the evolution proposal the paragraph is talking about. 

Done on mobile so might have missed something or this could be worded better? Use another PR or suggest changes for me to make - no problem either way.