In IPC article, use same string for port names in client and server
The name passed to `CFMessagePortCreateRemote` has to be the same name used when the port was created via `CFMessagePortCreateLocal` in the other process. This is a key part in how they figure out how to talk to each other. Otherwise `CFMessagePortCreateRemote` always returns `NULL`/`nil` since a port with the name given to it doesn't exist on the system. The example code might have tried to distinguish between the client and server by using different strings, but this causes the remote port creation to fail.