Update 2019-05-06-image-resizing.md
The current value is incorrect. It passes the number of bytes per row based on the size of the original image, not the size of the target (resized) image.

From docs:

```
bytesPerRow	
The number of bytes of memory to use per row of the bitmap. If the data parameter is NULL, passing a value of 0 causes the value to be calculated automatically.
```

I also ran the performance tests on iPhone 11 Pro running iOS 14.2 and they no longer match the results from the article. The fastest approach is #2 - based on `CGContext`. That matches my expectations. `UIGraphicsImageRendererContext` is simply a wrapper on top of `CGContext.` I don't expect it to be faster.