Remove use of method_exchangeImplementations (#770)
Setting a method's implementation directly is slow because the runtime
does not know which class it belongs to, so it clears out the entire
method cache on every class. If we know the class (and we do here) we
can use class_replaceMethod instead for similar effect, which will only
invalidate the class's cache.